+++ packages/tools/plugin-verifier/vitest.config.mjs
@@ -1,0 +1,11 @@
+// @ts-check
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    coverage: {
+      exclude: ['tests/fixtures/**'],
+    },
+    globals: true,
+  },
+});